requirements.txt
python-dotenv==1.1.0
snowflake-connector-python
pandas
seaborn
matplotlib
jupyter